requirements.txt
python-telegram-bot[webhooks]==21.6
httpx>=0.28.1
